requirements.txt
mcp[cli]>=1.0.0
google-genai>=1.0.0
Pillow>=10.0.0
uvicorn>=0.30.0
